build: migrate to tsup for bundling
This replaces the custom build script using esbuild and chokidar with
tsup, which uses esbuild under the hood and offers out of the box watch
capabilities, preserving previous configuration file compilation behavior.
