+++ package.json
@@ -17,8 +17,8 @@
     "type": "module",
     "main": "index.js",
     "scripts": {
-        "build": "npm run clean && node scripts/generate-command-index.js && node scripts/generate-manifests.js && node scripts/build.js",
-        "build:release": "npm run clean && node scripts/generate-command-index.js && node scripts/generate-manifests.js --release && node scripts/build.js --minify",
+        "build": "npm run clean && node scripts/generate-command-index.js && node scripts/generate-manifests.js && tsup",
+        "build:release": "npm run clean && node scripts/generate-command-index.js && node scripts/generate-manifests.js --release && tsup --minify",
         "check-deps": "node scripts/check-dependencies.js",
         "check-types": "tsc --noEmit",
         "clean": "rimraf packs/behavior/scripts",
@@ -36,7 +36,7 @@
         "test:duplication": "jscpd src/ --min-lines 5 --min-tokens 40 --max-lines 1000 --ignore 'src/**/__tests__/**'",
         "validate": "npm run check-deps && npx mct validate default PRJINT,UNKJSON,pathlength,PATHLENGTH,BASEGAMEVER,CHKMANIF,WPACKREFS,JSON,STRICT,ENTITYTYPE,FORMATVER,ITEMTYPE --display-only",
         "validate:report": "npm run check-deps && npx mct validate default PRJINT,UNKJSON,pathlength,PATHLENGTH,BASEGAMEVER,CHKMANIF,WPACKREFS,JSON,STRICT,ENTITYTYPE,FORMATVER,ITEMTYPE -o validation_report",
-        "watch": "node scripts/generate-command-index.js && concurrently \"tsc -w\" \"node scripts/build.js --watch\""
+        "watch": "node scripts/generate-command-index.js && concurrently \"tsc -w\" \"tsup --watch\""
     },
     "simple-git-hooks": {
         "pre-commit": "npx lint-staged"
@@ -98,6 +98,7 @@
         "prettier-plugin-packagejson": "latest",
         "rimraf": "latest",
         "simple-git-hooks": "latest",
+        "tsup": "^8.5.1",
         "type-fest": "latest",
         "typescript": "latest",
         "typescript-eslint": "latest",
